+++ src/hooks.ts
@@ -18,7 +18,7 @@ import { defaultConfig } from './defaults'
 type GetBinderTypeFromDomTarget<T extends Partial<GestureConfig>> = T['domTarget'] extends object ? Fn : ReactEventHandlers
 
 export function useGesture<Config extends Partial<GestureConfig>>(
-  handlers: GestureHandlersPartial | Handler<Coordinates>,
+  handlers: GestureHandlersPartial,
   config?: Config
 ): (...args: any[]) => GetBinderTypeFromDomTarget<Config> {
   // the gesture controller will keep track of all gesture states
@@ -26,14 +26,14 @@ export function useGesture<Config extends Partial<GestureConfig>>(
 
   if (!gestureController.current) {
     // we initialize the gesture controller once
-    gestureController.current = new GestureController(getDerivedHandlers(handlers), getDerivedConfig(config))
+    gestureController.current = new GestureController(handlers, getDerivedConfig(config))
   }
 
   React.useEffect(() => {
     // every time handlers or config change, we let the gesture controller compute
     // them so that the gesture handlers functions are aware of the changes
     gestureController.current!.config = getDerivedConfig(config)
-    gestureController.current!.handlers = getDerivedHandlers(handlers)
+    gestureController.current!.handlers = handlers
   }, [handlers, config])
 
   // when the user component unmounts, we run our gesture controller clean function
@@ -51,17 +51,6 @@ export const useHover = (handler: Handler<Coordinates>, config?: Partial<Gesture
 export const useScroll = (handler: Handler<Coordinates>, config?: Partial<GestureConfig>) => useGesture({ onScroll: handler }, config)
 export const useWheel = (handler: Handler<Coordinates>, config?: Partial<GestureConfig>) => useGesture({ onWheel: handler }, config)
 export const usePinch = (handler: Handler<DistanceAngle>, config?: Partial<GestureConfig>) => useGesture({ onPinch: handler }, config)
-
-function getDerivedHandlers(handlers: GestureHandlersPartial | Handler<Coordinates>): GestureHandlersPartial {
-  if (typeof handlers === 'function') return { onDrag: handlers }
-
-  const { onAction, ...rest } = handlers
-  const derivedHandlers = rest as GestureHandlersPartial
-
-  if (onAction) derivedHandlers.onDrag = onAction
-
-  return derivedHandlers
-}
 
 function getDerivedConfig(config?: Partial<GestureConfig>): GestureConfig {
   const derivedConfig = { ...defaultConfig, ...config }